requirements.txt
flask==3.0.3
streamlit==1.36.0
akshare==1.14.0
requests==2.32.3
pandas==2.2.2
passlib==1.7.4
apscheduler==3.10.4
openpyxl==3.1.5
python-dotenv==1.0.1
flask-cors==4.0.1
gunicorn==22.0.0
gevent==24.2.1
pyjwt==2.8.0
